organize printy
put modules in a folder printy and changed __init__ and __main__ imports
